[chore] 품종 데이터 수정

+++ src/main/kotlin/com/example/fatpetserver/pet/entity/Breeds.kt
@@ -27,8 +27,8 @@ class Breeds(
     val sex: Sex,
 
     @Column
-    val avgWeightLow: Float,
+    val avgWeightLow: Int,
 
     @Column
-    val avgWeightHigh: Float,
+    val avgWeightHigh: Int,
 ) : IdEntity()
